refactor contract manager into modules
