+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -19,10 +19,10 @@
 
 ### **Phase 2: Major Violations (500-999 lines)**
 - **Total Contracts**: 65
-- **Completed**: 0
+- **Completed**: 1
 - **In Progress**: 0
-- **Available**: 65
-- **Progress**: 0%
+- **Available**: 64
+- **Progress**: 1.5%
 
 ### **Phase 3: Moderate Violations (301-499 lines)**
 - **Total Contracts**: 98
@@ -54,6 +54,18 @@
 - ✅ `workflow_cli.py` (223 lines)
 
 **Validation**: All smoke tests pass, V2 standards compliant
+
+---
+### **CONTRACT #021: Contract Manager** ✅ **COMPLETED**
+**Agent**: Agent-3 (Infrastructure & DevOps)
+**Completion Date**: 2025-08-23
+**Files**: `src/core/contract_manager.py` split into focused modules
+**Deliverables**:
+- ✅ `contract_models.py` (99 lines)
+- ✅ `assignment_engine.py` (161 lines)
+- ✅ `validation.py` (75 lines)
+
+**Validation**: Unit tests pass, modules under 200 lines
 
 ---
 
@@ -103,12 +115,6 @@
 #### **CONTRACT #020: Advanced Workflow Engine**
 **File**: `src/core/advanced_workflow_engine.py` (861 lines)
 **Target**: 4 modules (≤250 lines each)
-**Priority**: HIGH
-**Estimated Duration**: 6 hours
-
-#### **CONTRACT #021: Contract Manager**
-**File**: `src/core/contract_manager.py` (711 lines)
-**Target**: 4 modules (≤200 lines each)
 **Priority**: HIGH
 **Estimated Duration**: 6 hours
