+++ examples/demo_core_agent_management_integration.py
@@ -23,7 +23,8 @@ sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
 
 from core.agent_registration_manager import AgentRegistrationManager
 from core.status_manager import StatusManager
-from core.contract_manager import ContractManager, ContractPriority, ContractStatus
+from core.assignment_engine import ContractManager
+from core.contract_models import ContractPriority, ContractStatus
 from core.performance_tracker import PerformanceTracker
 from core.agent_manager import AgentManager, AgentStatus, AgentCapability
 from core.config_manager import ConfigManager
